Adjust some of the layout and logging logic
