+++ auth/src/util/logger.go
@@ -1,0 +1,29 @@
+package util
+
+import (
+	"log"
+	"os"
+	"path/filepath"
+)
+
+var (
+    WarningLogger *log.Logger
+    InfoLogger    *log.Logger
+    ErrorLogger   *log.Logger
+)
+
+func init() {
+	ex, err := os.Executable()
+	if err != nil {
+		log.Fatalln(err)
+	}
+	logPath := filepath.Join(filepath.Dir(filepath.Dir(ex)), "auth.log")
+    file, err := os.OpenFile(logPath, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0666)
+    if err != nil {
+        log.Fatal(err)
+    }
+
+    InfoLogger = log.New(file, "INFO: ", log.Ldate|log.Ltime|log.Lshortfile)
+    WarningLogger = log.New(file, "WARNING: ", log.Ldate|log.Ltime|log.Lshortfile)
+    ErrorLogger = log.New(file, "ERROR: ", log.Ldate|log.Ltime|log.Lshortfile)
+}
